improve async init with connect middleware
