+++ src/resolvers.ts
@@ -1,12 +1,20 @@
 import * as path from 'path';
-import { RequestHandler } from "express";
-import { RouteMetadata } from "./framework/openapi.spec.loader";
+import { RequestHandler } from 'express';
+import { RouteMetadata } from './framework/openapi.spec.loader';
 
-export function defaultResolver(handlersPath: string, route: RouteMetadata): RequestHandler {
+const cache = {};
+export function defaultResolver(
+  handlersPath: string,
+  route: RouteMetadata,
+): RequestHandler {
   const tmpModules = {};
   const { expressRoute, method, schema } = route;
   const oId = schema['x-eov-operation-id'] || schema['operationId'];
   const baseName = schema['x-eov-operation-handler'];
+
+  const cacheKey = `${expressRoute}-${method}-${oId}-${baseName}`;
+  if (cache[cacheKey]) return cache[cacheKey];
+
   if (oId && !baseName) {
     throw Error(
       `found x-eov-operation-id for route ${method} - ${expressRoute}]. x-eov-operation-handler required.`,
@@ -17,11 +25,7 @@ export function defaultResolver(handlersPath: string, route: RouteMetadata): Req
       `found x-eov-operation-handler for route [${method} - ${expressRoute}]. operationId or x-eov-operation-id required.`,
     );
   }
-  if (
-    oId &&
-    baseName &&
-    typeof handlersPath === 'string'
-  ) {
+  if (oId && baseName && typeof handlersPath === 'string') {
     const modulePath = path.join(handlersPath, baseName);
     if (!tmpModules[modulePath]) {
       tmpModules[modulePath] = require(modulePath);
@@ -35,21 +39,27 @@ export function defaultResolver(handlersPath: string, route: RouteMetadata): Req
         `Could not find 'x-eov-operation-handler' with id ${oId} in module '${modulePath}'. Make sure operation '${oId}' defined in your API spec exists as a handler function in '${modulePath}'.`,
       );
     }
-    return tmpModules[modulePath][oId];
+
+    const handler = tmpModules[modulePath][oId];
+    cache[cacheKey] = handler;
+    return handler;
   }
 }
 
-export function modulePathResolver(handlersPath: string, route: RouteMetadata): RequestHandler {
-  const [controller, method] = route.schema['operationId'].split('.')
+export function modulePathResolver(
+  handlersPath: string,
+  route: RouteMetadata,
+): RequestHandler {
+  const [controller, method] = route.schema['operationId'].split('.');
 
   const modulePath = path.join(handlersPath, controller);
-  const handler = require(modulePath)
+  const handler = require(modulePath);
 
   if (handler[method] === undefined) {
     throw new Error(
-      `Could not find a [${method}] function in ${modulePath} when trying to route [${route.method} ${route.expressRoute}].`
-    )
+      `Could not find a [${method}] function in ${modulePath} when trying to route [${route.method} ${route.expressRoute}].`,
+    );
   }
 
-  return handler[method]
+  return handler[method];
 }
